Fixed Autoseacrh query and update to database
There was a bug in the code that outerjoin with NOL_AUTO_SEARCH_RESULTS was wrong. This resulted in not picking up new NOL rows. This fix has removed outer join and would make sure that a row is inserted to NOL_AUTOSEARCH_RESULTS is inserted if not already there