Style changes - define acronyms on first use
The first sentence of this readme is very acronym heavy, which makes
it hard to understand what this project does if you don't already
know the background.

I've rewritten this slightly to use the full names for NOLASA and MLRA
the first time they're introduced, and to link to the MLRA repo.