LASB-63 Changed scheduler to run at 6pm SUN through THU
NOLASA service overview meeting was held with HMCTS and CGI. Their feedback was that instead of running the application daily at 8pm it is better to run at 6pm. Moreover Fridays are not suitable due to potential release day. It was then agreed to run the scheduler Sunday through Thursday at 6pm.

Also changed the correxct characters for quotes