Allow AWS credentials to be provided in the environment
This is part of the configuration for shipping metrics to cloudwatch.

By default the library seems to look for profiles on the filesystem,
which doesn't work with docker. However the AWS client library ships
with a DefaultAWSCredentialProviderChain which seems to do the right
thing. See https://github.com/spring-cloud/spring-cloud-aws/issues/102

The environment variables needed are AWS_ACCESS_KEY_ID and
AWS_SECRET_ACCESS_KEY.

Note that when running locally, we need to pass these environment
variables to docker-compose.