Move measurement of the number of infox search results
We only look at the top 15 results, but it's useful to know if there
are more than 15, so count it right after getting the response.

Also add description metadata to metrics. You can also tag metrics, but
as far as I can tell cloudwatch doesn't do anything with those tags so
I've not added them.