Handle case where infox returns SUCCESS with zero results
I don't think we are currently handling this case correctly.

InfoXSearchStatus.FAILURE is used in these cases:
- we throw an exception building the request
- InfoX returns a status of LIBRA_FAILED_EXCEPTION or LIBRA_INVALID_CODE

Both of these cases should be counted as errors in the metrics, not zero
results.

As far as I can tell, if there are zero matches from Libra, then the
status will be SUCCESS, but we will create a InfoXSearchResult with an
array of nulls. So this is the case we need to return
ReconciliationResult.NO_MATCHES for.