Refactored code replaced Array with List
The application originally used arrays to store the search result which was mapped to numbered columns of the database. This unnecessarily required lot of boilerplate code. This became more visible when metrics functionality was added.

This has been simplified by providing view of List instead of individual columns from model object. Ideally the database table should have been normalized, however, making this change will ensure that the mapping to individual columns on the table (libraid1, libraid2, ....) will not be exposed to rest of the application.