Model chganges to match refactored data model
Data model related to NOLASA was not normalized. This was normalized via https://github.com/ministryofjustice/laa-mlra-application/pull/126. The changes made in this PR are related to this data model change. Model entities are mapped with the relational tables which required some changes in service and test classes.

This must be deployed alongwith https://github.com/ministryofjustice/laa-mlra-application/pull/126.