Don't send metrics every minute
By default, metrics are exported to cloudwatch every minute,
but the application only does anything when the scheduler runs, so this
results in a lot of unneccessary requests and clutters up the logs.

I was originally planning to set this to hourly, but I think that will
be a bit annoying because if we're running NOLASA on the test
environment, it could take up to 2 hours before we see anything on the
dashboard.

I've made it every 10m instead.

This doesn't change what the dashboard displays, because the dashboard
already summarises what happened each hour, due to this configuration:

```
{ "stat": "Sum", "period": 3600 }
```