LASB-110: 'Not found' is not correct
Metrics was reporting counts for errors instead of no match found, so fixed the logic to correctly pass the counts based on status returned.

Removed the logic of limiting the result set to 15. After normalizing the database there is no restrictions on how many search result we can save. However, as the MLRA user interface only shows 15 results, so it is good idea to provide this figure in metrics. So added results.over_15_count to metrics