Replaced Spring WS with Apache cxf for consuming Infox SOAP
Spring WS implmentation of JAXB caused few problems with Java 11 module system, so had to replace with Apache CXF.

There is no standard plugin to generate sources based on XSD / WSDL, currently using no.nils.wsdl2java. This works fine, however had to add large number of dependencies to support this in Java 11. This could be replaced in future with better plugin.

Currently there are few problems around SOAP version and XML name issue which we fixed for JAXB but needs to be fixed for cxf implementation.