Set version id for SOAP client to match with Infox service
There was an issue whereby the soap request message was being rejected by infox service. This was due to SOAP version difference.
Also added logging for payloads. Response message could be quite large so setting the limit of message size of 1k when logging.

Also cleaned up some code which was not required due cxf change.